src/TheftPreventionSystem/requirements.txt
picamera==1.13
RPi.GPIO==0.7.1
